fix: set default theme to avoid inconsistency and render "dark" mode per default when not not set
