+++ app/ChakraUIWrapper.tsx
@@ -2,8 +2,6 @@
 
 import { Provider } from 'components/ui/provider';
 
-//import { customTheme } from 'components/theme';
-
 export default function ChakraUIWrapper({ children }) {
-    return <Provider>{children}</Provider>;
+    return <Provider defaultTheme="dark">{children}</Provider>;
 }
